+++ packages/providers/lib/alchemy-provider.js
@@ -69,7 +69,7 @@ var AlchemyProvider = /** @class */ (function (_super) {
                 host = "eth-mainnet.alchemyapi.io/v2/";
                 break;
             case "goerli":
-                host = "eth-goerli.g.alchemyapi.io/v2/";
+                host = "eth-goerli.g.alchemy.com/v2/";
                 break;
             case "matic":
                 host = "polygon-mainnet.g.alchemy.com/v2/";